+++ async-raft/src/core/client.rs
@@ -320,7 +320,7 @@ impl<'a, D: AppData, R: AppDataResponse, N: RaftNetwork<D>, S: RaftStorage<D, R>
 
         let res = match resp {
             Ok(data) => {
-                let membership = if let EntryPayload::ConfigChange(ref c) = entry.payload {
+                let membership = if let EntryPayload::Membership(ref c) = entry.payload {
                     Some(c.membership.clone())
                 } else {
                     None
@@ -347,7 +347,7 @@ impl<'a, D: AppData, R: AppDataResponse, N: RaftNetwork<D>, S: RaftStorage<D, R>
 
     pub fn handle_special_log(&mut self, entry: &Entry<D>) {
         match &entry.payload {
-            EntryPayload::ConfigChange(ref mem) => {
+            EntryPayload::Membership(ref mem) => {
                 let m = &mem.membership;
                 if m.is_in_joint_consensus() {
                     // nothing to do